tests/requirements.txt
pytest==8.4.1
python-liquid==2.0.2